# go-backend/docs/swagger.yaml
openapi: 3.0.1
info:
  title: MessagerLoadAPI
  description: This is a API to practise scalling a messanger.
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:3000
tags:
  - name: general
    description: Endpoints without entities
  - name: conversation
    description: Management of conversations
  - name: user
    description: Management of users
paths:
  /health:
    get:
      tags:
        - general
      summary: Returns applications health
      operationId: getHealth
      responses:
        '200':
          description: List of movies in json format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'
  /api/v1/conversations/{id}:
    get:
      security:
        - BearerAuth: []
      tags:
        - conversation
      summary: Get a conversation
      operationId: getConversation
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier for a conversation
      responses:
        '200':
          description: Conversation in JSON format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '404':
          description: Conversation not found
  /api/v1/conversations/{id}/messages:
    get:
      security:
        - BearerAuth: []
      tags:
        - conversation
      summary: Get a conversation messages
      operationId: getConversationMessages
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier for a conversation
      responses:
        '200':
          description: Conversation in JSON format
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '404':
          description: Conversation not found
  /api/v1/conversations:
    post:
      security:
        - BearerAuth: []
      tags:
        - conversation
      summary: Creates a conversation
      operationId: createConversation
      responses:
        '200':
          description: Conversation created in JSON format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '500':
          description: Error creating conversation
  /api/v1/users:
    post:
      security:
        - BearerAuth: []
      tags:
        - user
      summary: Creates a user
      operationId: createUser
      requestBody:
        description: Create user request details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestBody'
      responses:
        '200':
          description: Created user in JSON format
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Error creating user
components:
  schemas:
    HealthCheck:
      type: object
      properties:
        status:
          type: string
          nullable: false
        imdb_data_downloading:
          type: boolean
          nullable: true
        building_db:
          type: boolean
          nullable: true
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from_user_id:
          type: string
        created_at:
          type: string
          format: datetime
        body:
          type: string
    CreateUserRequestBody:
      type: object
      properties:
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    ErrBadRequest:
      required:
        - status
        - title
        - detail
        - type
      properties:
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          example: 400
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Bad Request
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: There was something wrong with the request you made
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7231#section-6.6.4
          description: A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when dereferenced, it provide human-readable documentation for the problem type. When this member is not present, its value is assumed to be "about:blank".
    ErrUnprocessableEntity:
      required:
        - status
        - title
        - detail
        - type
      properties:
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          example: 422
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Unprocessabe Entity
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: The request was unprocessable
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7231#section-6.6.4
          description: A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when dereferenced, it provide human-readable documentation for the problem type. When this member is not present, its value is assumed to be "about:blank".
    ErrNotFound:
      required:
        - status
        - title
        - detail
        - type
      properties:
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          example: 404
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Record not found
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: The record you are looking for does not exist
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7231#section-6.6.4
          description: A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when dereferenced, it provide human-readable documentation for the problem type. When this member is not present, its value is assumed to be "about:blank".


  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer